Bug fixes
projectile functionality about 30% done
