+++ Gra/Gra/Gra.vcxproj
@@ -137,7 +137,6 @@
   <ItemGroup>
     <ClInclude Include="Header.h" />
     <ClInclude Include="LevelManager.h" />
-    <ClInclude Include="Projectile.h" />
     <ClInclude Include="Walls.h" />
     <ClInclude Include="Tanks.h" />
   </ItemGroup>
